Move test and check for pattern
