+++ pyscriptjs/tests/integration/test_02_output.py
@@ -1,5 +1,9 @@
+import base64
 import html
+import io
 import re
+
+from PIL import Image
 
 from .support import PyScriptTest
 
@@ -336,3 +340,33 @@ class TestOutput(PyScriptTest):
         console_text = self.console.all.lines
         assert console_text.index("1print") == (console_text.index("2print") - 1)
         assert console_text.index("1console") == (console_text.index("2console") - 1)
+
+    def test_image_renders_correctly(self):
+        """This is just a sanity check to make sure that images are rendered correctly."""
+        buffer = io.BytesIO()
+        img = Image.new("RGB", (4, 4), color=(0, 0, 0))
+        img.save(buffer, format="PNG")
+
+        b64_img = base64.b64encode(buffer.getvalue()).decode("utf-8")
+        expected_img_src = f"data:image/png;charset=utf-8;base64,{b64_img}"
+
+        self.pyscript_run(
+            """
+            <py-config>
+                packages = ["pillow"]
+            </py-config>
+
+            <div id="img-target" />
+            <py-script>
+                from PIL import Image
+                img = Image.new("RGB", (4, 4), color=(0, 0, 0))
+                display(img, target='img-target', append=False)
+            </py-script>
+            """
+        )
+
+        # TODO: This seems to be a py-script tag, should it?
+        rendered_img_src = self.page.locator("#py-internal-0 > img").get_attribute(
+            "src"
+        )
+        assert rendered_img_src == expected_img_src
